Seperated Patcher and Mod System PathfinderHacknet.exe produced from patcher PathfinderHacknet calls upon PathfinderHacknet.dll Fixed up lib Note: Must place AlienFXManagedWrapper3.5.dll, FNA.dll, and Steamworks.NET.dll into lib directory, not included for sake of license and game
