+++ HacknetPathfinder.sln
@@ -4,7 +4,7 @@ VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HacknetPathfinder", "HacknetPathfinder.csproj", "{B46D689F-4F1A-432C-8755-15BD50755016}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pathfinder", "..\Pathfinder\Pathfinder.csproj", "{56CC3A7F-445F-4121-9EF6-6C6C1B4F6001}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PathfinderPatcher", "PathfinderPatcher\PathfinderPatcher.csproj", "{FCD7FD47-6B38-47F4-9761-FC4290110125}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,10 +16,10 @@ Global
 		{B46D689F-4F1A-432C-8755-15BD50755016}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B46D689F-4F1A-432C-8755-15BD50755016}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B46D689F-4F1A-432C-8755-15BD50755016}.Release|Any CPU.Build.0 = Release|Any CPU
-		{56CC3A7F-445F-4121-9EF6-6C6C1B4F6001}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{56CC3A7F-445F-4121-9EF6-6C6C1B4F6001}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{56CC3A7F-445F-4121-9EF6-6C6C1B4F6001}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{56CC3A7F-445F-4121-9EF6-6C6C1B4F6001}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FCD7FD47-6B38-47F4-9761-FC4290110125}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FCD7FD47-6B38-47F4-9761-FC4290110125}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FCD7FD47-6B38-47F4-9761-FC4290110125}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FCD7FD47-6B38-47F4-9761-FC4290110125}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
