--- PathfinderModloader/Pathfinder.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace PathfinderModloader
-{
-    // This will be the modloader.
-    public class Pathfinder
-    {
-    }
-}
