+++ README.md
@@ -1,2 +1,5 @@
 # Hacknet-Pathfinder
 Hacknet Labyrinths Modloader in C#
+
+# Note
+Must include Hacknet's AlienFXManagedWrapper3.5.dll, FNA.dll, and Steamworks.NET.dll in the lib directory
