+++ lib/.gitignore
@@ -1,0 +1,4 @@
+AlienFXManagedWrapper3.5.dll
+FNA.dll
+Steamworks.NET.dll
+PathfinderPatcher.exe
